pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gregg-limper"
version = "0.1.0"
description = "Discord assistant for knowledge retrieval and response generation."
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "beautifulsoup4==4.13.4",
    "discord.py==2.5.2",
    "numpy==1.26.4",
    "ollama==0.5.3",
    "openai==1.97.1",
    "Pillow==11.3.0",
    "python-dotenv==1.0.1",
    "pymilvus==2.6.0",
    "PyNaCl==1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.1",
    "pytest-asyncio==0.24.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
